+++ packages/EasyCodingStandard/src/DependencyInjection/EasyCodingStandardKernel.php
@@ -7,7 +7,6 @@ use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\Config\Loader\LoaderResolver;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\DependencyInjection\Loader\DirectoryLoader;
 use Symfony\Component\DependencyInjection\Loader\GlobFileLoader;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
 use Symfony\Component\HttpKernel\Config\FileLocator;
@@ -57,9 +56,7 @@ final class EasyCodingStandardKernel extends AbstractCliKernel
      */
     public function registerBundles(): array
     {
-        return [
-            new CheckersBundle(),
-        ];
+        return [];
     }
 
     public function bootWithConfig(string $config): void
@@ -90,15 +87,13 @@ final class EasyCodingStandardKernel extends AbstractCliKernel
     /**
      * @param ContainerInterface|ContainerBuilder $container
      */
-    protected function getContainerLoader(ContainerInterface $container): DelegatingLoader
+    protected function getContainerLoader(ContainerInterface $container): LoaderInterface
     {
         $fileLocator = new FileLocator($this);
-        $loaderResolver = new LoaderResolver([
-            new CheckerTolerantYamlFileLoader($container, $fileLocator),
-            new GlobFileLoader($container, $fileLocator),
-            new DirectoryLoader($container, $fileLocator),
-        ]);
 
-        return new DelegatingLoader($loaderResolver);
+        return new DelegatingLoader(new LoaderResolver([
+            new GlobFileLoader($container, $fileLocator),
+            new CheckerTolerantYamlFileLoader($container, $fileLocator),
+        ]));
     }
 }